requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
pycryptodome==3.21.0
PyPDF2==3.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
